+++ Lib/ntpath/pure.py
@@ -1,333 +1,16 @@
 # Module 'ntpath.pure' -- pure operations on WinNT/Win95 pathnames
 """Pure pathname manipulations, WindowsNT/95 version."""
-
-
 import os
-import sys
-import genericpath
-from genericpath import commonprefix
+import ntpath
+from ntpath import (altsep, basename, commonpath, commonprefix, curdir,
+                    dirname, extsep, isabs, isreserved, join, normcase,
+                    normpath, pardir, pathsep, sep, split, splitdrive,
+                    splitext, splitroot)
 
 __all__ = ["altsep", "basename", "commonpath", "commonprefix", "curdir",
            "dirname", "extsep", "isabs", "isreserved", "join", "normcase",
            "normpath", "pardir", "pathsep", "relpath", "sep", "split",
            "splitdrive", "splitext", "splitroot"]
-
-curdir = "."
-pardir = ".."
-extsep = "."
-sep = "\\"
-pathsep = ";"
-altsep = "/"
-
-def _get_bothseps(path):
-    if isinstance(path, bytes):
-        return b'\\/'
-    else:
-        return '\\/'
-
-# Normalize the case of a pathname and map slashes to backslashes.
-# Other normalizations (such as optimizing '../' away) are not done
-# (this is done by normpath).
-
-try:
-    from _winapi import (
-        LCMapStringEx as _LCMapStringEx,
-        LOCALE_NAME_INVARIANT as _LOCALE_NAME_INVARIANT,
-        LCMAP_LOWERCASE as _LCMAP_LOWERCASE)
-
-    def normcase(s):
-        """Normalize case of pathname.
-
-        Makes all characters lowercase and all slashes into backslashes.
-        """
-        s = os.fspath(s)
-        if not s:
-            return s
-        if isinstance(s, bytes):
-            encoding = sys.getfilesystemencoding()
-            s = s.decode(encoding, 'surrogateescape').replace('/', '\\')
-            s = _LCMapStringEx(_LOCALE_NAME_INVARIANT,
-                               _LCMAP_LOWERCASE, s)
-            return s.encode(encoding, 'surrogateescape')
-        else:
-            return _LCMapStringEx(_LOCALE_NAME_INVARIANT,
-                                  _LCMAP_LOWERCASE,
-                                  s.replace('/', '\\'))
-except ImportError:
-    def normcase(s):
-        """Normalize case of pathname.
-
-        Makes all characters lowercase and all slashes into backslashes.
-        """
-        s = os.fspath(s)
-        if isinstance(s, bytes):
-            return os.fsencode(os.fsdecode(s).replace('/', '\\').lower())
-        return s.replace('/', '\\').lower()
-
-
-def isabs(s):
-    """Test whether a path is absolute"""
-    s = os.fspath(s)
-    if isinstance(s, bytes):
-        sep = b'\\'
-        altsep = b'/'
-        colon_sep = b':\\'
-        double_sep = b'\\\\'
-    else:
-        sep = '\\'
-        altsep = '/'
-        colon_sep = ':\\'
-        double_sep = '\\\\'
-    s = s[:3].replace(altsep, sep)
-    # Absolute: UNC, device, and paths with a drive and root.
-    return s.startswith(colon_sep, 1) or s.startswith(double_sep)
-
-
-# Join two (or more) paths.
-def join(path, *paths):
-    path = os.fspath(path)
-    if isinstance(path, bytes):
-        sep = b'\\'
-        seps = b'\\/'
-        colon_seps = b':\\/'
-    else:
-        sep = '\\'
-        seps = '\\/'
-        colon_seps = ':\\/'
-    try:
-        result_drive, result_root, result_path = splitroot(path)
-        for p in paths:
-            p_drive, p_root, p_path = splitroot(p)
-            if p_root:
-                # Second path is absolute
-                if p_drive or not result_drive:
-                    result_drive = p_drive
-                result_root = p_root
-                result_path = p_path
-                continue
-            elif p_drive and p_drive != result_drive:
-                if p_drive.lower() != result_drive.lower():
-                    # Different drives => ignore the first path entirely
-                    result_drive = p_drive
-                    result_root = p_root
-                    result_path = p_path
-                    continue
-                # Same drive in different case
-                result_drive = p_drive
-            # Second path is relative to the first
-            if result_path and result_path[-1] not in seps:
-                result_path = result_path + sep
-            result_path = result_path + p_path
-        ## add separator between UNC and non-absolute path
-        if (result_path and not result_root and
-            result_drive and result_drive[-1] not in colon_seps):
-            return result_drive + sep + result_path
-        return result_drive + result_root + result_path
-    except (TypeError, AttributeError, BytesWarning):
-        genericpath._check_arg_types('join', path, *paths)
-        raise
-
-
-# Split a path in a drive specification (a drive letter followed by a
-# colon) and the path specification.
-# It is always true that drivespec + pathspec == p
-def splitdrive(p):
-    """Split a pathname into drive/UNC sharepoint and relative path specifiers.
-    Returns a 2-tuple (drive_or_unc, path); either part may be empty.
-
-    If you assign
-        result = splitdrive(p)
-    It is always true that:
-        result[0] + result[1] == p
-
-    If the path contained a drive letter, drive_or_unc will contain everything
-    up to and including the colon.  e.g. splitdrive("c:/dir") returns ("c:", "/dir")
-
-    If the path contained a UNC path, the drive_or_unc will contain the host name
-    and share up to but not including the fourth directory separator character.
-    e.g. splitdrive("//host/computer/dir") returns ("//host/computer", "/dir")
-
-    Paths cannot contain both a drive letter and a UNC path.
-
-    """
-    drive, root, tail = splitroot(p)
-    return drive, root + tail
-
-
-try:
-    from nt import _path_splitroot_ex as splitroot
-except ImportError:
-    def splitroot(p):
-        """Split a pathname into drive, root and tail.
-
-        The tail contains anything after the root."""
-        p = os.fspath(p)
-        if isinstance(p, bytes):
-            sep = b'\\'
-            altsep = b'/'
-            colon = b':'
-            unc_prefix = b'\\\\?\\UNC\\'
-            empty = b''
-        else:
-            sep = '\\'
-            altsep = '/'
-            colon = ':'
-            unc_prefix = '\\\\?\\UNC\\'
-            empty = ''
-        normp = p.replace(altsep, sep)
-        if normp[:1] == sep:
-            if normp[1:2] == sep:
-                # UNC drives, e.g. \\server\share or \\?\UNC\server\share
-                # Device drives, e.g. \\.\device or \\?\device
-                start = 8 if normp[:8].upper() == unc_prefix else 2
-                index = normp.find(sep, start)
-                if index == -1:
-                    return p, empty, empty
-                index2 = normp.find(sep, index + 1)
-                if index2 == -1:
-                    return p, empty, empty
-                return p[:index2], p[index2:index2 + 1], p[index2 + 1:]
-            else:
-                # Relative path with root, e.g. \Windows
-                return empty, p[:1], p[1:]
-        elif normp[1:2] == colon:
-            if normp[2:3] == sep:
-                # Absolute drive-letter path, e.g. X:\Windows
-                return p[:2], p[2:3], p[3:]
-            else:
-                # Relative path with drive, e.g. X:Windows
-                return p[:2], empty, p[2:]
-        else:
-            # Relative path, e.g. Windows
-            return empty, empty, p
-
-
-# Split a path in head (everything up to the last '/') and tail (the
-# rest).  After the trailing '/' is stripped, the invariant
-# join(head, tail) == p holds.
-# The resulting head won't end in '/' unless it is the root.
-
-def split(p):
-    """Split a pathname.
-
-    Return tuple (head, tail) where tail is everything after the final slash.
-    Either part may be empty."""
-    p = os.fspath(p)
-    seps = _get_bothseps(p)
-    d, r, p = splitroot(p)
-    # set i to index beyond p's last slash
-    i = len(p)
-    while i and p[i-1] not in seps:
-        i -= 1
-    head, tail = p[:i], p[i:]  # now tail has no slashes
-    return d + r + head.rstrip(seps), tail
-
-
-# Split a path in root and extension.
-# The extension is everything starting at the last dot in the last
-# pathname component; the root is everything before that.
-# It is always true that root + ext == p.
-
-def splitext(p):
-    p = os.fspath(p)
-    if isinstance(p, bytes):
-        return genericpath._splitext(p, b'\\', b'/', b'.')
-    else:
-        return genericpath._splitext(p, '\\', '/', '.')
-splitext.__doc__ = genericpath._splitext.__doc__
-
-
-# Return the tail (basename) part of a path.
-
-def basename(p):
-    """Returns the final component of a pathname"""
-    return split(p)[1]
-
-
-# Return the head (dirname) part of a path.
-
-def dirname(p):
-    """Returns the directory component of a pathname"""
-    return split(p)[0]
-
-
-_reserved_chars = frozenset(
-    {chr(i) for i in range(32)} |
-    {'"', '*', ':', '<', '>', '?', '|', '/', '\\'}
-)
-
-_reserved_names = frozenset(
-    {'CON', 'PRN', 'AUX', 'NUL', 'CONIN$', 'CONOUT$'} |
-    {f'COM{c}' for c in '123456789\xb9\xb2\xb3'} |
-    {f'LPT{c}' for c in '123456789\xb9\xb2\xb3'}
-)
-
-def isreserved(path):
-    """Return true if the pathname is reserved by the system."""
-    # Refer to "Naming Files, Paths, and Namespaces":
-    # https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file
-    path = os.fsdecode(splitroot(path)[2]).replace(altsep, sep)
-    return any(_isreservedname(name) for name in reversed(path.split(sep)))
-
-def _isreservedname(name):
-    """Return true if the filename is reserved by the system."""
-    # Trailing dots and spaces are reserved.
-    if name[-1:] in ('.', ' '):
-        return name not in ('.', '..')
-    # Wildcards, separators, colon, and pipe (*?"<>/\:|) are reserved.
-    # ASCII control characters (0-31) are reserved.
-    # Colon is reserved for file streams (e.g. "name:stream[:type]").
-    if _reserved_chars.intersection(name):
-        return True
-    # DOS device names are reserved (e.g. "nul" or "nul .txt"). The rules
-    # are complex and vary across Windows versions. On the side of
-    # caution, return True for names that may not be reserved.
-    return name.partition('.')[0].rstrip(' ').upper() in _reserved_names
-
-
-# Normalize a path, e.g. A//B, A/./B and A/foo/../B all become A\B.
-# Previously, this function also truncated pathnames to 8+3 format,
-# but as this module is called "ntpath", that's obviously wrong!
-try:
-    from nt import _path_normpath as normpath
-
-except ImportError:
-    def normpath(path):
-        """Normalize path, eliminating double slashes, etc."""
-        path = os.fspath(path)
-        if isinstance(path, bytes):
-            sep = b'\\'
-            altsep = b'/'
-            curdir = b'.'
-            pardir = b'..'
-        else:
-            sep = '\\'
-            altsep = '/'
-            curdir = '.'
-            pardir = '..'
-        path = path.replace(altsep, sep)
-        drive, root, path = splitroot(path)
-        prefix = drive + root
-        comps = path.split(sep)
-        i = 0
-        while i < len(comps):
-            if not comps[i] or comps[i] == curdir:
-                del comps[i]
-            elif comps[i] == pardir:
-                if i > 0 and comps[i-1] != pardir:
-                    del comps[i-1:i+1]
-                    i -= 1
-                elif i == 0 and root:
-                    del comps[i]
-                else:
-                    i += 1
-            else:
-                i += 1
-        # If the path is now empty, substitute '.'
-        if not prefix and not comps:
-            comps.append(curdir)
-        return prefix + sep.join(comps)
-
 
 def relpath(path, start):
     """Return a relative version of a path"""
@@ -336,96 +19,4 @@ def relpath(path, start):
     if not isabs(path) or not isabs(start):
         raise ValueError("paths are not absolute")
 
-    if isinstance(path, bytes):
-        sep = b'\\'
-        curdir = b'.'
-        pardir = b'..'
-    else:
-        sep = '\\'
-        curdir = '.'
-        pardir = '..'
-
-    try:
-        start_drive, _, start_rest = splitroot(start)
-        path_drive, _, path_rest = splitroot(path)
-        if normcase(start_drive) != normcase(path_drive):
-            raise ValueError("path is on mount %r, start on mount %r" % (
-                path_drive, start_drive))
-
-        start_list = start_rest.split(sep) if start_rest else []
-        path_list = path_rest.split(sep) if path_rest else []
-        # Work out how much of the filepath is shared by start and path.
-        i = 0
-        for e1, e2 in zip(start_list, path_list):
-            if normcase(e1) != normcase(e2):
-                break
-            i += 1
-
-        rel_list = [pardir] * (len(start_list)-i) + path_list[i:]
-        if not rel_list:
-            return curdir
-        return sep.join(rel_list)
-    except (TypeError, ValueError, AttributeError, BytesWarning, DeprecationWarning):
-        genericpath._check_arg_types('relpath', path, start)
-        raise
-
-
-# Return the longest common sub-path of the iterable of paths given as input.
-# The function is case-insensitive and 'separator-insensitive', i.e. if the
-# only difference between two paths is the use of '\' versus '/' as separator,
-# they are deemed to be equal.
-#
-# However, the returned path will have the standard '\' separator (even if the
-# given paths had the alternative '/' separator) and will have the case of the
-# first path given in the iterable. Additionally, any trailing separator is
-# stripped from the returned path.
-
-def commonpath(paths):
-    """Given an iterable of path names, returns the longest common sub-path."""
-    paths = tuple(map(os.fspath, paths))
-    if not paths:
-        raise ValueError('commonpath() arg is an empty iterable')
-
-    if isinstance(paths[0], bytes):
-        sep = b'\\'
-        altsep = b'/'
-        curdir = b'.'
-    else:
-        sep = '\\'
-        altsep = '/'
-        curdir = '.'
-
-    try:
-        drivesplits = [splitroot(p.replace(altsep, sep).lower()) for p in paths]
-        split_paths = [p.split(sep) for d, r, p in drivesplits]
-
-        # Check that all drive letters or UNC paths match. The check is made only
-        # now otherwise type errors for mixing strings and bytes would not be
-        # caught.
-        if len({d for d, r, p in drivesplits}) != 1:
-            raise ValueError("Paths don't have the same drive")
-
-        drive, root, path = splitroot(paths[0].replace(altsep, sep))
-        if len({r for d, r, p in drivesplits}) != 1:
-            if drive:
-                raise ValueError("Can't mix absolute and relative paths")
-            else:
-                raise ValueError("Can't mix rooted and not-rooted paths")
-
-        common = path.split(sep)
-        common = [c for c in common if c and c != curdir]
-
-        split_paths = [[c for c in s if c and c != curdir] for s in split_paths]
-        s1 = min(split_paths)
-        s2 = max(split_paths)
-        for i, c in enumerate(s1):
-            if c != s2[i]:
-                common = common[:i]
-                break
-        else:
-            common = common[:len(s1)]
-
-        return drive + root + sep.join(common)
-    except (TypeError, AttributeError):
-        genericpath._check_arg_types('commonpath', *paths)
-        raise
+    return ntpath.relpath(path, start)
